requirements.txt
fast_hdbscan==0.1.3
hdbscan==0.8.33
matplotlib==3.8.4
numpy==1.26.4
pandas==1.5.3
scikit_learn==1.4.2
sentence_transformers==2.6.1
umap_learn==0.5.6
